+++ src/app/view-config-upgraders.js
@@ -339,14 +339,37 @@ export function upgradeFrom1_0_8(config) {
 
 // Added in version 1.0.11:
 // - Changes to spatial layer coordination type names.
+// - Cell -> Obs, Gene -> Feature in coordination type names.
 export function upgradeFrom1_0_10(config) {
   const coordinationSpace = { ...config.coordinationSpace };
 
   const scopeAnalogies = {
+    // Spatial layer types
     spatialRasterLayers: 'spatialRasterLayer',
     spatialCellsLayer: 'spatialSegmentationLayer',
     spatialMoleculesLayer: 'spatialPointLayer',
     spatialNeighborhoodsLayer: 'spatialNeighborhoodLayer',
+    // Other types
+    cellFilter: 'obsFilter',
+    cellHighlight: 'obsHighlight',
+    cellSelection: 'obsSelection',
+    cellSetSelection: 'obsSetSelection',
+    cellSetHighlight: 'obsSetHighlight',
+    cellSetColor: 'obsSetColor',
+    geneFilter: 'featureFilter',
+    geneHighlight: 'featureHighlight',
+    geneSelection: 'featureSelection',
+    geneExpressionColormap: 'featureValueColormap',
+    geneExpressionColormapRange: 'featureValueColormapRange',
+    cellColorEncoding: 'obsColorEncoding',
+    additionalCellSets: 'additionalObsSets',
+    embeddingCellSetPolygonsVisible: 'embeddingObsSetPolygonsVisible',
+    embeddingCellSetLabelsVisible: 'embeddingObsSetLabelsVisible',
+    embeddingCellSetLabelSize: 'embeddingObsSetLabelSize',
+    embeddingCellRadius: 'embeddingObsRadius',
+    embeddingCellRadiusMode: 'embeddingObsRadiusMode',
+    embeddingCellOpacity: 'embeddingObsOpacity',
+    embeddingCellOpacityMode: 'embeddingObsOpacityMode',
   };
 
   Object.entries(scopeAnalogies).forEach(([oldKey, newKey]) => {